requirements.txt
discord.py==2.1.0
python-decouple==3.7
requests==2.28.2
